+++ scripts/client.js
